requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
av==11.0.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cmake==3.28.1
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
cython==3.0.7
dataclasses-json==0.6.3
decord==0.6.0
distro==1.8.0
einops==0.4.1
exceptiongroup==1.2.0
fairscale==0.4.4
filelock==3.13.1
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==4.4.3
fvcore==0.1.5.post20221221
gradio==4.22.0
gradio-client==0.13.0
# gensim==3.8.3
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.19.4
idna==3.6
imageio==2.33.1
importlib-resources==6.1.1
# install==1.3.5
iopath==0.1.10
jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.2
langchain-community==0.0.14
langchain-core==0.1.14
langchain-openai==0.0.3
langchainhub==0.1.14
langcodes==3.3.0
langsmith==0.0.83
lapx==0.5.5
lit==17.0.6
markupsafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
moviepy==1.0.3
networkx==3.2.1
nltk==3.8.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.9.0
opencv-python==4.8.1.78
packaging==23.2
pandas==1.3.5
parameterized==0.9.0
pathy==0.10.3
pillow==10.1.0
pims==0.6.1
portalocker==2.8.2
preshed==3.0.9
protobuf==4.21.12
psutil==5.9.7
py-cpuinfo==9.0.0
pydantic==2.5.3
pyparsing==3.1.1
python-dateutil==2.8.2
pytorchvideo==0.1.5
pytz==2023.3.post1
pyyaml==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.1
scikit-learn
scipy==1.11.4
seaborn==0.13.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
slicerator==1.1.0
smart-open==6.4.0
sniffio==1.3.0
sqlalchemy==2.0.25
srsly==2.4.8
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
# theano
thinc==8.1.12
thop==0.1.1-2209072238
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.5.4
tokenizers==0.12.1
torch
torchvision
tqdm==4.66.1
transformers==4.27.0
triton==2.1.0
types-requests==2.31.0.20240106
typing-extensions==4.9.0
typing-inspect==0.9.0
ultralytics==8.0.235
urllib3==2.1.0
wasabi==0.10.1
wcwidth==0.2.12
webdataset==0.2.86
webencodings==0.5.1
xdg==6.0.0
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0
git+https://github.com/openai/CLIP.git
git+https://github.com/Maluuba/nlg-eval.git@master
